refactor: change export type in firebase/index.ts and exports values in AppRouter.tsx
